[libwspcodec] Repackaging + content type fix
Moved the sources to wspcodec directory so that it resembles the
layout of the installed include files. Makes it easier to build
something that references libwspcodec without installing the devel
package.

Made content type field configurable. This resolves the conflict
between WSP Content-Type and X-Mms-Report-Allowed (both are 0x11)

Bumped version to 2.0.0 because this update is not backward compatible.